requirements.txt
asgiref==3.8.1
click==7.1.2
click-didyoumean==0.3.1
Django==5.1.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
drf-yasg==1.21.7
inflection==0.5.1
p==1.5.0
packaging==24.1
psycopg2-binary==2.9.10
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
setuptools==75.2.0
sqlparse==0.5.1
uritemplate==4.1.1
